Version 20200108.1
http://www.gurux.fi/node/14348

+++ development/src/GXDLMSImageTransfer.cpp
@@ -217,7 +217,7 @@ int CGXDLMSImageTransfer::ImageBlockTransfer(CGXDLMSClient* client, CGXByteBuffe
     std::vector<CGXByteBuffer> blocks;
     if ((ret = GetImageBlocks(image, blocks)) == 0)
     {
-        imageBlockCount = blocks.size();
+        imageBlockCount = (unsigned long) blocks.size();
         CGXDLMSVariant tmp;
         for (std::vector<CGXByteBuffer>::iterator it = blocks.begin(); it != blocks.end(); ++it)
         {
